refactor single & multi query mapping logic
